mehnatai-backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1

# Utilities
python-multipart==0.0.19
